+++ assets/js/migration.js
@@ -7,7 +7,7 @@
  * @package SilverAssist\ContactFormToAPI
  * @subpackage Assets
  * @since 1.3.4
- * @version 1.3.4
+ * @version 1.3.5
  * @author Silver Assist
  */