+++ assets/js/settings-page.js
@@ -5,7 +5,7 @@
  *
  * @package SilverAssist\ContactFormToAPI
  * @since 1.2.0
- * @version 1.3.4
+ * @version 1.3.5
  */
 
 (function ($) {