+++ includes/Admin/RequestLogTable.php
@@ -8,7 +8,7 @@
  * @package SilverAssist\ContactFormToAPI
  * @subpackage Admin
  * @since 1.1.0
- * @version 1.3.4
+ * @version 1.3.5
  * @author Silver Assist
  */