+++ includes/Services/ExportService.php
@@ -8,7 +8,7 @@
  * @package SilverAssist\ContactFormToAPI
  * @subpackage Services
  * @since 1.2.0
- * @version 1.3.4
+ * @version 1.3.5
  * @author Silver Assist
  */