+++ includes/Utils/DateFilterTrait.php
@@ -7,7 +7,7 @@
  *
  * @package SilverAssist\ContactFormToAPI\Utils
  * @since   1.2.0
- * @version 1.3.4
+ * @version 1.3.5
  * @author  Silver Assist
  */